requirements.txt
asgiref==3.6.0
Django==4.2.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-taggit==4.0.0
Pillow==9.5.0
psycopg2-binary==2.9.6
sqlparse==0.4.4
